+++ Tests/SentryTests/Integrations/SentryOutOfMemoryTrackerTests.swift
@@ -24,7 +24,7 @@ class SentryOutOfMemoryTrackerTests: XCTestCase {
             
             crashWrapper = TestSentryCrashAdapter.sharedInstance()
             
-            let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: crashWrapper)
+            let hub = SentryHub(client: client, andScope: nil, andCrashAdapter: crashWrapper, andCurrentDateProvider: currentDate)
             SentrySDK.setCurrentHub(hub)
             
             fileManager = try! SentryFileManager(options: options, andCurrentDateProvider: currentDate)
@@ -51,6 +51,8 @@ class SentryOutOfMemoryTrackerTests: XCTestCase {
         super.tearDown()
         sut.stop()
         fixture.fileManager.deleteAllFolders()
+        
+        clearTestState()
     }
 
     func testStart_StoresAppState() {