+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## unreleased
+## Unreleased
 
+- feat: Add automatic http request performance monitoring (#1178)
 - feat: Add clearAttachments to Scope (#1195)
 
 ## 7.2.0-beta.2